+++ src/utils.rs
@@ -48,6 +48,11 @@ pub fn parse_fasta_dataset(dataset: &str) -> RosalindResult<Vec<String>> {
     Ok(buf.split(delimiter).map(|s| s.to_string()).collect())
 }
 
+/// This calculates factorial value of given `n`
+pub fn factorial(n: usize) -> usize {
+    if n == 0 || n == 1 { 1 } else { (2..n+1).fold(1, |p, i| p * i) }
+}
+
 #[test]
 fn it_should_determine_fasta_label() {
     assert_eq!(is_fasta_label(">Rosalind_1"), true);
@@ -56,7 +61,7 @@ fn it_should_determine_fasta_label() {
 
 #[cfg(test)]
 mod tests {
-    use super::parse_fasta_dataset;
+    use super::{parse_fasta_dataset, factorial};
 
     #[test]
     fn it_should_parse_fasta_dataset() {
@@ -79,5 +84,17 @@ mod tests {
         let dataset = parse_fasta_dataset(fasta_dataset).unwrap();
 
         assert_eq!(dataset, expected_dataset);
+    }
+
+    #[test]
+    fn it_should_return_factrorial_of_n() {
+        assert_eq!(factorial(0), 1);
+        assert_eq!(factorial(1), 1);
+        assert_eq!(factorial(2), 2);
+        assert_eq!(factorial(3), 6);
+        assert_eq!(factorial(4), 24);
+        assert_eq!(factorial(5), 120);
+        assert_eq!(factorial(6), 720);
+        assert_eq!(factorial(7), 5040);
     }
 }